refactor: move Node into its own module
